pyproject.toml
[tool.poetry]
name = "chocolate"
version = "0.1.0"
description = ""
authors = ["llc <86468766@qq.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.13,<4.0"
uvicorn = "^0.35.0"
langchain = "^1.1.3"
langchain-core = "^1.1.2"
requests = "^2.32.5"
langchain-google-genai = "^2.1.10"
langchain-openai = "^0.3.32"
pydantic = "^2.11.7"
fastapi = "^0.124.0"
starlette = "^0.50.0"
langchain-community = "^0.3.29"
poetry-core = "^2.1.3"
chromadb = "^1.0.20"
pytest = "^8.4.1"
python-multipart = "^0.0.20"
easyocr = "^1.7.2"
paddlepaddle = "^3.1.1"
paddleocr = "^3.2.0"
pytesseract = "^0.3.13"
opencv-python = "^4.12.0.88"
ffmpeg-python = "^0.2.0"
openai-whisper = "^20250625"
speechrecognition = "^3.14.3"
pydub = "^0.25.1"
python-logging-loki = "^0.3.1"
sentence-transformers = "^5.1.0"
transformers = "^4.56.0"
matplotlib = "^3.10.6"
einops = "^0.8.1"
transformers-stream-generator = "^0.0.5"
accelerate = "^1.10.1"
optimum = "^1.27.0"
sentencepiece = "^0.2.1"
sacremoses = "^0.1.1"
watchdog = "^6.0.0"
jieba = "^0.42.1"
meilisearch = "^0.38.0"
pymupdf = "^1.26.6"
python-docx = "^1.2.0"
pandas = "^2.3.3"
openpyxl = "^3.1.5"
xlrd = "^2.0.2"
langchain-text-splitters = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
